+++ www/src/routes/_app/create.tsx
@@ -12,6 +12,8 @@ import {
 } from '@/modules/create/preset/storage'
 import { PreviewPanel } from '@/modules/create/preview/preview-panel'
 import { StudioPanel } from '@/modules/create/studio'
+import { StudioShellProvider } from '@/modules/create/studio/shell'
+import { BuilderTopBar } from '@/modules/create/studio/top-bar'
 
 type MobilePane = 'customize' | 'preview'
 
@@ -67,29 +69,38 @@ function CreatePage() {
   }, [designSystem])
 
   return (
-    <div className="flex h-[calc(100svh-var(--header-height))] min-h-0 flex-1 flex-col gap-3 p-4 pt-2 lg:flex-row lg:gap-6 lg:p-6 lg:pt-2">
-      {/* Mobile-only view switcher — hidden once the two panes fit side by side. */}
-      <ToggleButtonGroup
-        aria-label="Editor view"
-        selectionMode="single"
-        disallowEmptySelection
-        size="sm"
-        selectedKeys={[mobilePane]}
-        onSelectionChange={(keys) => {
-          const next = keys.values().next().value
-          if (next === 'customize' || next === 'preview') setMobilePane(next)
-        }}
-        className="w-full shrink-0 *:flex-1 lg:hidden"
-      >
-        <ToggleButton id="customize">Customize</ToggleButton>
-        <ToggleButton id="preview">Preview</ToggleButton>
-      </ToggleButtonGroup>
-      <StudioPanel
-        className={cn(mobilePane === 'preview' && 'max-lg:hidden')}
-      />
-      <PreviewPanel
-        className={cn(mobilePane === 'customize' && 'max-lg:hidden')}
-      />
-    </div>
+    // The builder owns its own chrome on /create: the global site header is
+    // skipped for this route (see routes/_app/route.tsx) and stood in for by
+    // BuilderTopBar, so the two never stack into a double bar.
+    <StudioShellProvider>
+      <div className="flex flex-col">
+        <BuilderTopBar />
+        <div className="flex h-[calc(100svh-var(--header-height))] min-h-0 flex-1 flex-col gap-3 p-4 pt-2 lg:flex-row lg:gap-6 lg:p-6 lg:pt-2">
+          {/* Mobile-only view switcher — hidden once the two panes fit side by side. */}
+          <ToggleButtonGroup
+            aria-label="Editor view"
+            selectionMode="single"
+            disallowEmptySelection
+            size="sm"
+            selectedKeys={[mobilePane]}
+            onSelectionChange={(keys) => {
+              const next = keys.values().next().value
+              if (next === 'customize' || next === 'preview')
+                setMobilePane(next)
+            }}
+            className="w-full shrink-0 *:flex-1 lg:hidden"
+          >
+            <ToggleButton id="customize">Customize</ToggleButton>
+            <ToggleButton id="preview">Preview</ToggleButton>
+          </ToggleButtonGroup>
+          <StudioPanel
+            className={cn(mobilePane === 'preview' && 'max-lg:hidden')}
+          />
+          <PreviewPanel
+            className={cn(mobilePane === 'customize' && 'max-lg:hidden')}
+          />
+        </div>
+      </div>
+    </StudioShellProvider>
   )
 }
